finish

+++ src/modules/customer/customer.service.ts
@@ -1,29 +1,52 @@
 import { pool } from "../../config/db"
 
 
-const getUser = async()=>{
-const result = await pool.query(`
+const getUser = async () => {
+  
+    const result = await pool.query(`
     SELECT * FROM users
     `)
     return result
 }
 
-const updateUser =async( name:string,email:string,password:string,phone:string,id:string )=>{
-     
- const result =  await pool.query(`
-    UPDATE users SET name=$1,email=$2,password=$3,phone=$4 WHERE id=$5 RETURNING *
-    `,[name,email,password,phone,id])
-    return result;
+const updateUser = async (name: string, role: string, tokenID: string, id: string) => {
+ 
+    if(role === "admin"){
+       const result = await pool.query(`
+       UPDATE users SET name=$1 WHERE id=$2 RETURNING *
+     `, [name, id])
+     return result
+    }
+
+    if(role=== "customer" && tokenID === id){
+        const update = await pool.query(`
+             UPDATE users SET name=$1 WHERE id=$2 RETURNING *
+            `,[name,id]);
+          return update  
+    }
+    throw new Error("Unauthorizes");
+
+  
+
+
+
 }
 
-const deleteUser =async(id:string)=>{
-      const result = await pool.query(`
+const deleteUser = async (id: string) => {
+const  bookingStatus  = await pool.query(`
+    SELECT status FROM bookings WHERE id=$1
+    `,[id]);
+    const isActiveBooking=bookingStatus.rows[0].status;
+    if(isActiveBooking === 'active'){
+        throw new Error("user booking is active")
+    }
+    const result = await pool.query(`
     DELETE FROM users WHERE id=$1
-    `,[id])
+    `, [id])
     return result
 }
 
-export const customerServices ={
+export const customerServices = {
     getUser,
     updateUser,
     deleteUser
